Harden user level permissions to controls route + refactor
* Refactored backend logic for get_commands() function so it checks users
  permissions and factors that into final commands list.
  - If a user doesn't have permissions for some control it just wont even be
    included in the list.
  - So now if a user tries to run a command they don't have access to, its not
    just hidden on the frontend. Now it won't even exist in the list of
    available cmds for the user on the backend.
* Make sure originally added admin user (user with id = 1) cannot be deleted.
* Plus misc tweaks to improve readablity & make more pythonic.
  - I'm going to go through and do this to everything at some point anyways.
    Might as well start now with newly written functions.

+++ app/auth.py
@@ -149,6 +149,10 @@ def edit_users():
                 flash("Cannot delete currently logged in user!", category='error')
                 return redirect(url_for('auth.edit_users'))
 
+            if user_ident.id == 1:
+                flash("Cannot delete main admin user!", category='error')
+                return redirect(url_for('auth.edit_users'))
+
             db.session.delete(user_ident)
             db.session.commit()
             flash(f"User {selected_user} deleted!")
@@ -191,7 +195,6 @@ def edit_users():
             if user_ident == None:
                 flash("Invalid user selected!", category='error')
                 return redirect(url_for('auth.edit_users'))
-
 
         permissions = dict()
         role = 'user' # Default to user role.
